==> 00 exploration.ipynb <==
import pandas as pd
import numpy as np

df=pd.read_csv("../data/raw/survey_results_responses.csv")

df.shape

df.sample(1).iloc[0]

df.info()

df.describe()

questions=['YearsCode','YearsCodePro']

for col in questions :
    print(col)
    print(df[col].unique())
    print("------------------------------")
    print( )
