==> 01 preprocessing.ipynb <==
import pandas as pd
import numpy as np

df=pd.read_csv("../data/raw/survey_results_responses.csv")

def split_answer(data_series,delimiter=";"):
    
    def is_splittable(pd_series,delimiter):
        return pd_series.str.contains(delimiter)
    
    def splitting(pd_series,delimiter):
        return pd_series.str.split(delimiter)
    
    check_splittable= is_splittable(data_series,delimiter)
    
    if not check_splittable.any() :
        
        return data_series
    
    modefied_ans= splitting(data_series,delimiter)
    
    mask_null=modefied_ans.isnull()
    
    modefied_ans.loc[mask_null]=modefied_ans.loc[mask_null].apply(lambda x : [])
    
    return modefied_ans

modefied_df=df.copy()

replace_dict={
    "YearsCodePro": {'Less than 1 year': 0 , 'More than 50 years' : 51},
    "YearsCode": {'Less than 1 year': 0 , 'More than 50 years' : 51}
}

for col , replacement in replace_dict.items() :
    
    modefied_df[col]=modefied_df[col].replace(replacement).astype(np.float32)

object_cols=modefied_df.select_dtypes(include='object').columns.tolist()

for col in object_cols :
    df[col]=split_answer(df[col])

i=df.sample(1).index[0]

print(df['LanguageHaveWorkedWith'].iloc[i])
print("---------------------------")
print(df['WebframeHaveWorkedWith'].iloc[i])

df.to_pickle("../data/processed/1_processed_df.pkl")
